Fix RSSI issue and remove dead code
Removed commented code
fixed issue with RSSI not showing. Need to set connect to true

+++ Feather/Feather9x_TX_M0/Feather9x_TX_M0.ino
@@ -214,7 +214,7 @@ void loop()
       Serial.println((char*)buf);
       Serial.print("RSSI: ");
       Serial.println(rf95.lastRssi(), DEC); 
-      oled.setRSSI(rf95.lastRssi());   
+      oled.setRSSI((int)rf95.lastRssi());   
     }
     else
     {
